code/requirements.txt
six
tqdm==4.30.*
torch==1.0.1
git+https://github.com/pytorch/text.git@master#wheel=torchtext
future
configargparse
PyYAML
tensorflow==1.15.0
tensorboardX
py-rouge
transformers
simpletransformers
sklearn
pandas
numpy
OpenNMT-py
git+https://github.com/pytorch/fairseq
